* Overhauled lighting (exponential instead of linear behavior)  * Changed the way daylight is calculated  * Changed the way the day-night cycle is calculated  * The day-night cycle is now more fluent  * Fixed a bug where trees did not cast shadows correctly

+++ src/com/github/begla/blockmania/Main.java
@@ -412,7 +412,7 @@ public final class Main {
                 }
             } else if (parsingResult.get(0).equals("set")) {
                 if (parsingResult.get(1).equals("time")) {
-                    _world.setTime(Short.parseShort(parsingResult.get(2)));
+                    _world.setTime(Float.parseFloat(parsingResult.get(2)));
                     success = true;
                     // Otherwise try lookup the given variable within the settings
                 } else {
